refactor(simulator): replace redundant CreateLockableScooter refactor(connector/common): encapsulate createMessageHandlerWithTimeout logic
