+++ connector/common/src/vehicle/model/builders/create/lockableScooter/CreateLockableScooter.ts
@@ -99,3 +99,5 @@ export class CreateLockableScooter {
     );
   }
 }
+
+export default CreateLockableScooter;
